+++ CHANGES.md
@@ -4,7 +4,7 @@ New language:
 
 - *MikroTik* RouterOS Scripting language by [Ivan Dementev][].
 
-New style:
+New styles:
 
 - *VisualStudio 2015 Dark* by [Nicolas LLOBERA][]
 - *Lightfair* by [Tristian Kelly][]